TIP-856: Move Pim\Console class to Akeneo\Tool
