+++ phpspec.yml.dist
@@ -30,8 +30,8 @@ suites:
         spec_path: src/Akeneo/Component/Classification
         src_path: src/Akeneo/Component/Classification
     Console:
-        namespace: Akeneo\Component\Console
-        psr4_prefix: Akeneo\Component\Console
+        namespace: Akeneo\Tool\Component\Console
+        psr4_prefix: Akeneo\Tool\Component\Console
         spec_path: src/Akeneo/Component/Console
         src_path: src/Akeneo/Component/Console
     FileStorage:
