+++ src/Akeneo/Tool/Component/Console/CommandResult.php
@@ -4,7 +4,7 @@
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
-namespace Akeneo\Component\Console;
+namespace Akeneo\Tool\Component\Console;
 
 /**
  * Command result object
